chore: add documents to the db only after successful payment
