+++ api/controllers/CheckUploadedDocumentsController.js
@@ -1,6 +1,5 @@
 const sails = require('sails');
 
-const inDevEnvironment = process.env.NODE_ENV === 'development';
 const COST_PER_DOCUMENT = 30;
 
 const CheckUploadedDocumentsController = {
@@ -13,55 +12,14 @@ const CheckUploadedDocumentsController = {
 
     addDocsToDBHandler(req, res) {
         try {
-            const { uploadedFileData } = req.session.eApp;
-            if (uploadedFileData.length === 0) {
-                throw new Error('No files uploaded');
-            }
-            for (let i = 0; i <= uploadedFileData.length; i++) {
-                const endOfLoop = i === uploadedFileData.length;
-                if (endOfLoop) {
-                    CheckUploadedDocumentsController._checkDocumentCountAndPaymentDetails(
-                        req,
-                        res
-                    );
-
-                    break;
-                }
-                UploadedDocumentUrls.create(
-                    CheckUploadedDocumentsController._dbColumnData(
-                        uploadedFileData[i],
-                        req
-                    )
-                )
-                    .then(() => {
-                        sails.log.info(
-                            `Url for document ${uploadedFileData[i].filename} added to db`
-                        );
-                    })
-                    .catch((err) => {
-                        throw new Error(err);
-                    });
-            }
+            CheckUploadedDocumentsController._checkDocumentCountAndPaymentDetails(
+                req,
+                res
+            );
         } catch (err) {
             sails.log.error(err.message);
             res.serverError();
         }
-    },
-
-    _dbColumnData(uploadedFile, req) {
-        const sessionData = req.session;
-        const fileUrl = inDevEnvironment
-            ? uploadedFile.storageName
-            : uploadedFile.location;
-
-        if (!sessionData.appId) {
-            throw new Error('Missing application id');
-        }
-        return {
-            application_id: sessionData.appId,
-            uploaded_url: fileUrl,
-            filename: uploadedFile.filename,
-        };
     },
 
     _checkDocumentCountAndPaymentDetails(req, res) {
